Fix: ReferenceError in JavaScript code preventing refresh
Const variables will be locally scoped to the block instead of scoped
to the global object. Therefore, when the if body for
`poke_reload_window`'s type executes, the `poke_reload_window`
will only be defined in that if body's block.
This means that the call in `poke_expire_check` will not succeed
as the function is not defined in that scope.
Hence a ReferenceError will occur and the page will not be refreshed.
Fix this by just replacing the call to `poke_reload_window` by its
contents.